Fixed player spawning sometimes upside-down.
The player would spawn upside-down if the
ship was upside-down. I replicated this bug
by using the physics gun to flip the ship,
then respawning in the ship. The issue is that
after you have spawned (directly after PlayerSpawn)
something causes the camera to flip to align
with the spawn. A timer is used so that the
player is flipped after they have spawned and
their screen has been flipped. To reset the camera,
ResetCamera() is called. This keeps the same
pitch and yaw, but resets roll to zero. If
the player is already right-side up, they will
not notice anything, but if a player spawns
upside down, 1/2 a second afterwards, they
will be flipped back up. I tested to confirmed
that it worked on two different maps.